+++ packages/react/src/components/select.tsx
@@ -1,5 +1,6 @@
 import * as React from "react";
 import { cx } from "../cx";
+import { Icon } from "./icon";
 
 export interface SelectOption {
   value: string;
@@ -93,9 +94,7 @@ export function Select({
         onKeyDown={onTriggerKeyDown}
       >
         <span>{selected ? selected.label : placeholder}</span>
-        <svg viewBox="0 0 16 16" width="16" height="16" fill="none" aria-hidden="true">
-          <path d="M4 6l4 4 4-4" stroke="currentColor" strokeWidth="1.5" strokeLinecap="round" strokeLinejoin="round" />
-        </svg>
+        <Icon name="chevron-right" rotate={90} />
       </button>
       {open && (
         <div id={`${idBase}-listbox`} role="listbox" className="rs-menu" aria-activedescendant={`${idBase}-opt-${activeIndex}`}>